init redis setup + publish message on droneTopic #2

+++ truck/src/main/java/com/polytech/si5/al/dronedelivery/team/g/truck/dto/DeliveryDto.java
@@ -1,6 +1,6 @@
 package com.polytech.si5.al.dronedelivery.team.g.truck.dto;
 
 public class DeliveryDto {
-    public Long latitude;
-    public Long longitude;
+    public double latitude;
+    public double longitude;
 }
